.github/scripts/release-notes/pyproject.toml
[project]
name = "release-notes"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pygithub>=2.0.0",
    "packaging>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-datadir>=1.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
